Scale Poisson input by dt
lambda*dt should be the input of the Poisson process, but this was
not the case before. The maximum number of spikes per bin(of 10 ms)
was ~14 which is not the case in reality. Now the max is 3. This
change required scaling the fire rate output of NL estimation
methods as well.