Change eigenvector function to eigh to fix complex values
Eigenvectors and eigenvalues of symmetrical values (such as the
covariance matrix we are using) are always real; however numerical
errors can lead to small imaginary values in eigenvectors of
covariance matrix in STC analysis. To alleviate np.linalg.eigh is
now used; it assumes the matrix is symmetrical and always returns
real values.